refactoring

+++ CrudService/src/main/java/com/unipampa/crud/controller/AccommodationController.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.unipampa.crud.dto.AccommodationDTO;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +15,8 @@ import com.unipampa.crud.service.AccommodationService;
 import com.unipampa.crud.model.Accommodation;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/api/accommodations")
@@ -27,6 +31,9 @@ public class AccommodationController {
 
 	@PostMapping("/add")
 	@ApiOperation(value = "Salva uma acomodação)")
+	@ApiResponses(value = {
+			@ApiResponse(code = 200, message = "Recurso salvo com sucesso!" )
+	})
 	public ResponseEntity<Accommodation> save(@RequestBody AccommodationDTO accommodationDTO) {
 		var accommodation = mapper.convertValue(accommodationDTO, Accommodation.class);
 		accommodationService.save(accommodation);
@@ -39,17 +46,30 @@ public class AccommodationController {
 		List<Accommodation> accommodations = accommodationService.findAll();
 		return ResponseEntity.status(HttpStatus.OK).body(accommodations);
 	}
-	
+
 	@GetMapping("/find/{id}")
 	@ApiOperation(value = "Encontra uma acomodação através do id")
-	public ResponseEntity<Accommodation> getById(@PathVariable("id") Long id) {
-		Optional<Accommodation> hosting = accommodationService.findById(id);
-		return new ResponseEntity<>(hosting.get(), HttpStatus.OK);
+	@ApiResponses(value = {
+			@ApiResponse(code = 200, message = "Recurso encontrado com sucesso"),
+			@ApiResponse(code = 404, message = "Recurso não encontrado")
+	})
+	public ResponseEntity<Object> getById(@PathVariable("id") String id) {
+		Optional<Accommodation> accomodation = accommodationService.findById(id);
+		return accomodation.<ResponseEntity<Object>>map(accommodation -> new ResponseEntity<>(accommodation, HttpStatus.OK))
+				.orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).body("Acomodação não encontrada para esse id"));
 	}
 	
 	@DeleteMapping("/delete/{id}")
 	@ApiOperation(value = "Deleta uma acomodação através do id")
-	public ResponseEntity<Void> deleteAccommodation(@PathVariable("id") Long id) {
+	@ApiResponses(value = {
+			@ApiResponse(code = 200, message = "Recurso deletado com sucesso"),
+			@ApiResponse(code = 404, message = "Recurso não encontrado")
+	})
+	public ResponseEntity<Object> deleteAccommodation(@PathVariable("id") String id) {
+		Optional<Accommodation> accommodation = accommodationService.findById(id);
+		if(accommodation.isEmpty()){
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Acomodação não encontrada para esse id, portanto não pode ser deletada!");
+		}
 		accommodationService.delete(id);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
